README-covesa-cdsp-iotdb.md: minor - clarify readme scope
Add a minor note to clarify the CDSP readme is intended to cover the
extensions to the upstream project contained in this fork, i.e. help
people to understand how to use the Kuksa CAN Provider with CDSP.
For wider help look to the Provider documentation, e.g. README.md

Signed-off-by: Stephen Lawrence <stephen.lawrence@renesas.com>